Extract GameHandler
