+++ backend/src/test/kotlin/com/game/main/ServerTest.kt
@@ -10,11 +10,13 @@ import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import java.io.InputStream
 
 class ServerTest {
 
     private val wsHandlerMock = mockk<GameWebSocket>(relaxed = true)
-    private val app = gameServerHandler("src/test/resources/test-frontend-assets/", apiHandler(wsHandlerMock))
+    private val gameHandler = GameHandler()
+    private val app = gameServerHandler("src/test/resources/test-frontend-assets/", gameHandler.apiHandler(wsHandlerMock))
 
     @BeforeEach
     fun setup() {
@@ -30,6 +32,7 @@ class ServerTest {
     fun `serves files`() {
         assertEquals(Status.OK, get("/").status)
         assertEquals(Status.OK, get("/lobby/123").status)
+        assertEquals(Status.OK, get("/game/456").status)
         assertEquals(Status.OK, get("/anything").status)
     }
 
@@ -37,23 +40,18 @@ class ServerTest {
     fun `Game is updated whenever a new player joins`() {
         val gameId = "randomGameId"
         GameRepository.createGame(gameId, Game(gameId))
-
-        val username1 = "testUser1"
-        val username2 = "testUser2"
-        val request1 = Request(Method.POST, "api/game/$gameId/")
-            .body(Jackson.asInputStream(JoinGameRequest(gameId, username1)))
-        val request2 = Request(Method.POST, "api/game/$gameId/")
-            .body(Jackson.asInputStream(JoinGameRequest(gameId, username2)))
         val game = GameRepository.getGame(gameId)!!
 
-        joinGameHandler(request1, wsHandlerMock)
+        val username1 = "testUser1"
+        post("api/game/$gameId/join", Jackson.asInputStream(JoinGameRequest(gameId, username1)))
         val players1 = game.playerListForSerialization()
         assert(players1.size == 1)
         assert(game.countPlayers() == 1)
         assert(game.hostId == username1)
         assert(players1.contains(username1))
 
-        joinGameHandler(request2, wsHandlerMock)
+        val username2 = "testUser2"
+        post("api/game/$gameId/join", Jackson.asInputStream(JoinGameRequest(gameId, username2)))
         val players2 = game.playerListForSerialization()
         assert(players2.size == 2)
         assert(game.countPlayers() == 2)
@@ -64,13 +62,11 @@ class ServerTest {
 
     @Test
     fun `should return 404 if game not found`() {
-        val gameId = "randomGameId"
-        val requestBody = JoinGameRequest(gameId, "username")
-        val request = Request(Method.POST, "/join-game/$gameId")
-            .body(Jackson.asInputStream(requestBody))
-        val response = joinGameHandler(request, wsHandlerMock)
+        val gameId = "notFoundGameId"
+        val response = post("api/game/$gameId/join", Jackson.asInputStream(JoinGameRequest(gameId, "username")))
         assertEquals(Status.NOT_FOUND, response.status)
     }
 
     private fun get(path: String) = app(Request(Method.GET, path))
+    private fun post(path: String, body: InputStream) = app(Request(Method.POST, path).body(body))
 }
