+++ frontend/src/Landing.js
@@ -5,7 +5,7 @@ export default function Landing() {
   let versionLinkPart = version.replace("-snapshot", "")
   return (
     <div className="buttonbox">
-      <form action="/api/new-game" method="post">
+      <form action="/api/game/new" method="post">
         <button type="submit">Play online with friends</button>
       </form>
       <a href="/kit"><button>Dice and categories for playing in person</button></a>
